Move git repo check into utils
