+++ persephone/utils.py
@@ -1,15 +1,27 @@
 """ Miscellaneous utility functions. """
-
 import os
-
-import numpy as np
-
-from nltk.metrics import distance
-
 import subprocess
 from subprocess import PIPE
 
+from git import Repo
+import numpy as np
+from nltk.metrics import distance
+
 from . import config
+
+def is_git_directory_clean(path_to_repo, search_parent_dirs):
+    """
+    Check that the git working directory is in a clean state
+    and raise exceptions if not.
+    :path_to_repo: The path of the git repo
+    """
+    repo = Repo(path_to_repo, search_parent_dirs)
+    if repo.untracked_files:
+        raise DirtyRepoException("Untracked files. Commit them first.")
+    # If there are changes to already tracked files
+    if repo.is_dirty():
+        raise DirtyRepoException("Changes to the index or working tree."
+                                 "Commit them first .")
 
 def target_list_to_sparse_tensor(target_list):
     """ Make tensorflow SparseTensor from list of targets, with each element in
